fix: icons structure
