+++ src/components/icons/logo-flat/logo-flat.tsx
@@ -1,0 +1,32 @@
+import type { LogoFlatProps } from "./types";
+
+const LogoFlat = ({ size = 24, color = "currentColor" }: LogoFlatProps) => {
+  return (
+    <svg
+      xmlns="http://www.w3.org/2000/svg"
+      width={size}
+      height={size}
+      viewBox="0 0 120 200"
+      fill="none"
+    >
+      <path
+        d="M102.699 128C109.874 107.5 97.3659 89.5 93.6992 86C105.699 120 59.6995 140.5 28.1987 143C15.7284 143.99 11.1989 155.249 15.6992 165C24.6993 184.5 39.6992 181.5 49.6992 177C57.6992 173.4 66.0326 167.5 69.1992 165C88.6992 150 99.0242 138.5 102.699 128Z"
+        fill={color}
+      />
+      <path
+        d="M16.4446 71.4433C9.2944 91.9519 21.8245 109.937 25.4954 113.432C13.4541 79.4469 59.4292 58.8913 90.927 56.3532C103.396 55.3484 107.912 44.0832 103.4 34.3381C94.3761 14.849 79.3799 17.8671 69.3853 22.3792C61.3897 25.9889 53.0635 31.8989 49.8998 34.4028C30.418 49.4263 20.1069 60.9388 16.4446 71.4433Z"
+        fill={color}
+      />
+      <path
+        d="M62.6995 90.7621C79.6995 111.263 61.6995 130.263 45.1995 137.762C88.1995 126.263 93.6995 103.763 93.6995 94.7627C93.6995 84.2627 89.0009 79.1657 84.6995 76.7621C67.6995 67.2626 57.79 84.8393 62.6995 90.7621Z"
+        fill={color}
+      />
+      <path
+        d="M56.4927 108.197C39.3648 87.8034 57.2457 68.6912 73.6985 61.0888C30.7712 72.8568 25.412 95.3908 25.4683 104.391C25.5339 114.89 30.2642 119.958 34.5807 122.335C51.6397 131.728 61.4391 114.089 56.4927 108.197Z"
+        fill={color}
+      />
+    </svg>
+  );
+};
+
+export { LogoFlat };
