+++ src/components/icons/logo-flat/types.ts
@@ -1,0 +1,4 @@
+export interface LogoFlatProps {
+  size?: number;
+  color?: string;
+}
